backend/requirements.txt
face_recognition==1.3.0
opencv-python>=4.8.0,<4.9
numpy==1.24.3
faiss-cpu
groq
python-dotenv
langchain
langchain-groq
langchain-community
openai
tiktoken